backend/requirements.txt
Flask==3.0.2
python-escpos==3.1
Flask-Cors==4.0.0